reverts merge of PR #16
