--- lib/app/chats/domain/entities/chat_message.dart
@@ -1,19 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
-
-class ChatMessage extends Equatable {
-  final String id;
-  final String msg;
-  final String uid;
-  final DateTime timestamp;
-
-  ChatMessage({
-    @required this.id,
-    @required this.msg,
-    @required this.uid,
-    @required this.timestamp,
-  });
-
-  @override
-  List<Object> get props => [id, msg, uid, timestamp];
-}
