--- lib/core/error/exceptions.dart
@@ -1,3 +1,0 @@
-class ServerException implements Exception {}
-
-class CacheException implements Exception {}
