+++ lib/core/managers/navigation/locator.dart
@@ -1,12 +1,15 @@
-import 'package:get/get.dart';
 
-import '../../../app/auth/presentation/viewmodels/auth_page_model.dart';
-import '../../../app/auth/presentation/viewmodels/tabs_page_model.dart';
-import '../../../app/mood_analysis/presentation/viewmodels/mood_analysis_page_model.dart';
 
-void setupLocator() {
-  //register view models
-  Get.lazyPut(() => TabsPageModel());
-  Get.lazyPut(() => AuthPageModel());
-  Get.lazyPut(() => MoodAnalysisPageModel());
-}
+// import '../../../app/presentation/viewmodels/auth/auth_page_model.dart';
+// import '../../../app/presentation/viewmodels/mood_analysis/mood_analysis_page_model.dart';
+// import '../../../app/presentation/viewmodels/tabs_page_model.dart';
+// import '../../services/navigation_service.dart';
+
+// GetIt locator = GetIt.instance;
+
+// void setupLocator() {
+//   //register view models
+//   Get.lazyPut(() => TabsPageModel());
+//   Get.lazyPut(() => AuthPageModel());
+//   Get.lazyPut(() => MoodAnalysisPageModel());
+// }
