+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:maser/core/managers/navigation/locator.dart';
 
 import 'core/managers/camera_manager.dart';
@@ -11,14 +10,14 @@ import 'core/managers/theme/app_theme.dart';
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   cameras = await availableCameras();
-  setupLocator();
+  // setupLocator();
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
+    return MaterialApp(
       debugShowCheckedModeBanner: false,
       // navigatorKey: locator<NavigationService>().navigatorKey,
       title: 'Flutter Demo',
