+++ pubspec.yaml
@@ -33,7 +33,7 @@ dependencies:
   fluttertoast: ^8.0.7
   tflite: ^1.1.2
   flutter_login: ^2.0.0
-  equatable: ^2.0.3
+  get_it: ^7.1.3
 
 dev_dependencies:
   auto_route_generator: ^2.1.0
