--- test/fixtures/chat_group.json
@@ -1,18 +1,0 @@
-{
-    "id": "test id",
-    "mentorId": "test mentorId",
-    "messages": [
-        {   
-            "id": "test id",
-            "msg": "test msg",
-            "uid": "test uid",
-            "timestamp": "2000-01-01T00:00:00.000"
-        },
-        {   
-            "id": "test id",
-            "msg": "test msg",
-            "uid": "test uid",
-            "timestamp": "2000-01-01T00:00:00.000"
-        }
-    ]
-}
