--- test/fixtures/chat_message.json
@@ -1,6 +1,0 @@
-{   
-    "id": "test id",
-    "msg": "test msg",
-    "uid": "test uid",
-    "timestamp": "2000-01-01T00:00:00.000"
-}
