--- test/fixtures/fixture_reader.dart
@@ -1,3 +1,0 @@
-import 'dart:io';
-
-String fixture(String name) => File('test/fixtures/$name').readAsStringSync();
