--- test/fixtures/story.json
@@ -1,9 +1,0 @@
-{
-    "id": "test id",
-    "storyTitle": "test storyTitle",
-    "storyContent": "test storyContent",
-    "imageUrl": "test imageUrl",
-    "postedOn": "2000-01-01T00:00:00.000",
-    "mentorId": "test mentorId",
-    "likeCount": 1
-}
