--- test/fixtures/user.json
@@ -1,7 +1,0 @@
-{
-    "id" : "test id",
-    "name" : "test name",
-    "email" : "test email",
-    "phone" : "1234567890",
-    "imageUrl" : "test url"
-}
